backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
prisma==0.11.0